fix(build): fix schema generation issues (#63)
This fixes a few issues relating to the CI json schema generation.

- befcc1f452e2226d329b67b0510bd5f2e4a75591 fixes #60, and also fixes
several other servers that previously had an empty `properties` in the
schema. This happened because `c.properties` is a table, which means
`vim.list_extend` doesn't do anything. `properties` should also always
be an object and not an array, so `vim.empty_dict()` is now being used
to prevent outputting `[]` when `properties` is empty.
- 84e19f41b4c0ad6865fdc61f62bae5cff496b71f embeds `index.json` rather
than downloading it every time. As mentioned in #54, the gist is no
longer maintained, so there's no reason to download it anymore. In the
future this should probably be [sourced from
elsewhere](https://gist.github.com/williamboman/a01c3ce1884d4b57cc93422e7eae7702?permalink_comment_id=4689512).
- 2a911fd425dabc54847d61c387fabb4cd572907e fixes the URL for Volar.
- 6dfd503dc5a09a040dab00c2922d4d4e0f9374b5 is just me running the build
step locally. If I shouldn't have done this I'm happy to revert it.

Fixes #52
Fixes #60

Co-authored-by: Folke Lemaitre <folke.lemaitre@gmail.com>